# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Cumulocity Service Request API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "basicAuth" : [ ]
  } ],
  "paths" : {
    "/api/service/request/{serviceRequestId}" : {
      "get" : {
        "tags" : [ "service-request-controller" ],
        "summary" : "GET service request by Id",
        "description" : "Returns service request by internal Id",
        "operationId" : "getServiceRequestById",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequest"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequest"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "service-request-controller" ],
        "summary" : "PUT service request by Id",
        "description" : "Updates specific service request.",
        "operationId" : "updateServiceRequestById",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceRequestPatchRqBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequest"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequest"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "service-request-controller" ],
        "summary" : "DELETE service request by Id",
        "description" : "Delete a service request object at Cumulocity IoT. Related object at external system will not be deleted!",
        "operationId" : "deleteServiceRequestById",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    },
...
    "/api/service/request/comment/{commentId}" : {
      "put" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "PUT service request comment by Id",
        "description" : "updates specific service request comment. This operation is only allowed by owner of comment!",
        "operationId" : "patchServiceRequestCommentById",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "Internal comment Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceRequestCommentRqBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequestComment"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequestComment"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequestComment"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "DELETE service request comment by Id",
        "description" : "deletes specific service request comment. This operation is only allowed by owner of comment!",
        "operationId" : "deleteServiceRequestCommentById",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "Internal comment Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "404" : {
            "description" : "Not Found"
          },
          "403" : {
            "description" : "Forbidden"
          }
        }
      }
    },
...
    "/api/service/request" : {
      "get" : {
        "tags" : [ "service-request-controller" ],
        "summary" : "GET service request list",
        "description" : "Returns a list of all service requests in IoT Platform. Additional query parameters allow to filter that list. The default configuration will return all service requests which are not closed! With parameter all=true, all service requests will be returned without fillter.",
        "operationId" : "getServiceRequestList",
        "parameters" : [ {
          "name" : "sourceId",
          "in" : "query",
          "description" : "Filter, returns all service request equal source Id",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "statusList",
          "in" : "query",
          "description" : "Filter, returns all service request with status defined in that list",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "priorityList",
          "in" : "query",
          "description" : "Filter, returns all service request with priority defined in that list",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }, {
          "name" : "all",
          "in" : "query",
          "description" : "filter, \"true\" returns all service request, \"false\" (default) returns only active service requests.",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Indicates how many entries of the collection shall be returned. The upper limit for one page is 2,000 objects.",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "currentPage",
          "in" : "query",
          "description" : "The current page of the paginated results.",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "orderBy",
          "in" : "query",
          "description" : "OrderBy, orders list by status and/or priority and/or timestamp.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "withTotalPages",
          "in" : "query",
          "description" : "When set to true, the returned result will contain in the statistics object the total number of pages. Only applicable on range queries.",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "Filter, returns all service request of the type defined. Current supported types are: alarm, note, maintenance, downtime, other",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withSourceAssets",
          "in" : "query",
          "description" : "Filter, returns all service request with source assets",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "withSourceDevices",
          "in" : "query",
          "description" : "Filter, returns all service request with source devices",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestListServiceRequest"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "service-request-controller" ],
        "summary" : "CREATE service request",
        "description" : "Creates a new service request object at Cumulocity IoT Platform.",
        "operationId" : "createServiceRequest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceRequestPostRqBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "409" : {
            "description" : "Conflict, Alarm already assigned to service request!",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "412" : {
            "description" : "Precondition Failed, Alarm of service request not found!",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequest"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request, Service request body is invalid!",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/service/request/{serviceRequestId}/comment" : {
      "get" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "Returns all comments of specific service request by internal Id.",
        "description" : "Each service request can have n comments. This endpoint returns the complete list of comments of a specific service request.",
        "operationId" : "getServiceRequestCommentList",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Indicates how many entries of the collection shall be returned. The upper limit for one page is 2,000 objects.",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "currentPage",
          "in" : "query",
          "description" : "The current page of the paginated results.",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "withTotalPages",
          "in" : "query",
          "description" : "When set to true, the returned result will contain in the statistics object the total number of pages. Only applicable on range queries.",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ServiceRequestComment"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RequestListServiceRequestComment"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "Add new service request comment to specific service request.",
        "description" : "Each service request can have n comments. This endpoint adds a new comment to a specific service request.",
        "operationId" : "createServiceRequestComment",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceRequestCommentRqBody"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequestComment"
                }
              }
            }
          },
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ServiceRequestComment"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/service/request/comment/{commentId}/attachment" : {
      "get" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "DOWNLOAD attachment for specific comment",
        "description" : "Download attachment for comment",
        "operationId" : "downloadServiceRequestCommentAttachment",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "Internal comment Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "service-request-comment-controller" ],
        "summary" : "UPLOAD attachment for specific comment",
        "description" : "Upload attachment for service request comment",
        "operationId" : "uploadServiceRequestCommentAttachment",
        "parameters" : [ {
          "name" : "commentId",
          "in" : "path",
          "description" : "Internal comment Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "file",
          "in" : "query",
          "description" : "Mulitpart file, attachment",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "binary"
          }
        }, {
          "name" : "force",
          "in" : "query",
          "description" : "Controls if the attachment can be overwritten. force == true means file will be overwritten if exists, otherwise a http 409 will be returned.",
          "required" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "409" : {
            "description" : "Conflict"
          },
          "404" : {
            "description" : "Not Found"
          },
          "201" : {
            "description" : "Created"
          }
        }
      }
    },
...
    "/api/adapter/service/request/{serviceRequestId}/comment" : {
      "get" : {
        "tags" : [ "service-request-external-controller" ],
        "summary" : "Returns all user comments of specific service request by internal Id.",
        "description" : "Each service request can have n comments. This endpoint returns the complete list of user comments of a specific service request.",
        "operationId" : "getServiceRequestCommentList_1",
        "parameters" : [ {
          "name" : "serviceRequestId",
          "in" : "path",
          "description" : "Internal service request Id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ServiceRequestComment"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ServiceRequestComment"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ServiceRequestPatchRqBody" : {
        "type" : "object",
        "description" : "Service Request object at Cumulocity IoT Platform. This object is the representative / proxy for external system object like ticket, issue, notification, incident etc.. It contains the life cycle information of it, IoT data references (device, alarms, events and measurement) and meta data for the service request like owner, creation time etc.. It has a flexible key & value list \"customProperties\" to add custom properties as well.",
        "properties" : {
          "status" : {
            "$ref" : "#/components/schemas/ServiceRequestStatus",
            "description" : "Service request status"
          },
          "priority" : {
            "$ref" : "#/components/schemas/ServiceRequestPriority",
            "description" : "Servcie request priority"
          },
          "title" : {
            "type" : "string",
            "description" : "Service request title / summary"
          },
          "description" : {
            "type" : "string",
            "description" : "Service request detailed description"
          },
          "isActive" : {
            "type" : "boolean",
            "description" : "Service request active flag, shows if the service request is active!",
            "example" : true
          },
          "externalId" : {
            "type" : "string",
            "description" : "Service request external ID, contains the service request object ID of the external system.",
            "example" : 123456789
          },
          "order" : {
            "$ref" : "#/components/schemas/ServiceOrder",
            "description" : "Service Order"
          },
          "customProperties" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            },
            "description" : "Custom specific properties"
          }
        }
      },
...
      "ServiceRequestCommentRqBody" : {
        "type" : "object",
        "description" : "Comment object of service request. Service request object can have n comments.",
        "properties" : {
          "text" : {
            "type" : "string",
            "description" : "comment text",
            "example" : "This is my comment ..."
          },
          "externalId" : {
            "type" : "string",
            "description" : "Service request comment external ID, contains the service request comment object ID of the external system.",
            "example" : 123456789
          },
          "isClosed" : {
            "type" : "boolean",
            "description" : "Service request closed flag, shows if the service request comment is closed! This fragment is basicaly used for retention rules."
          },
          "type" : {
            "type" : "string",
            "description" : "comment type enumeration (USER, SYSTEM)",
            "enum" : [ "user", "system" ]
          },
          "owner" : {
            "type" : "string",
            "description" : "creator / owner ",
            "example" : "owner@example.com",
            "readOnly" : true
          },
          "attachment" : {
            "$ref" : "#/components/schemas/ServiceRequestAttachment",
            "description" : "File attachment of Service Request"
          }
        },
        "required" : [ "owner", "text", "type" ]
      },
...
      "ErrorResponseBody" : {
        "type" : "object",
        "description" : "Response body for error. This can contain additional error message for the client.",
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          }
        }
      },
...
      "RequestListServiceRequest" : {
        "type" : "object",
        "description" : "Request list",
        "properties" : {
          "list" : {
            "type" : "array",
            "description" : "list element",
            "items" : {
              "$ref" : "#/components/schemas/ServiceRequest"
            }
          },
          "currentPage" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "current page",
            "example" : 2
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "page size",
            "example" : 1
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "totalPages",
            "example" : 3
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "totalElements",
            "example" : 83
          }
        },
        "required" : [ "currentPage", "list", "pageSize", "totalPages" ]
      },
...
    },
    "securitySchemes" : {
      "basicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      }
    }
  }
}